# openapi.yaml
paths:
...
  /api/studio/episodes:
    post:
      summary: "Create an episode \u2014 chunk text and enqueue generat"
      description: "Create an episode \u2014 chunk text and enqueue generation."
      responses:
        '200':
          description: Success
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEpisodeBody'
    get:
      summary: List all episodes.
      description: List all episodes.
      responses:
        '200':
          description: Success
  /api/studio/episodes/{episode_id}:
    get:
      summary: Get episode with its chunks.
      description: Get episode with its chunks.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
    put:
      summary: Update episode metadata.
      description: Update episode metadata.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEpisodeBody'
    delete:
      summary: Delete episode and audio files.
      description: Delete episode and audio files.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/episodes/{episode_id}/regenerate:
    post:
      summary: Re-generate all chunks for an episode.
      description: Re-generate all chunks for an episode.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/episodes/{episode_id}/regenerate-with-settings:
    post:
      summary: Re-generate episode with new settings, with undo s
      description: Re-generate episode with new settings, with undo support.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegenerateWithSettingsBody'
  /api/studio/undo/{undo_id}:
    post:
      summary: Undo an episode regeneration within the grace peri
      description: Undo an episode regeneration within the grace period.
      responses:
        '200':
          description: Success
      parameters:
      - name: undo_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/episodes/bulk-move:
    post:
      summary: Move multiple episodes to a folder.
      description: Move multiple episodes to a folder.
      responses:
        '200':
          description: Success
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkMoveEpisodesBody'
  /api/studio/episodes/bulk-delete:
    post:
      summary: Delete multiple episodes.
      description: Delete multiple episodes.
      responses:
        '200':
          description: Success
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteEpisodesBody'
  /api/studio/episodes/{episode_id}/chunks/{chunk_index}/regenerate:
    post:
      summary: Regenerate a single chunk.
      description: Regenerate a single chunk.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
      - name: chunk_index
        in: path
        required: true
        schema:
          type: integer
  /api/studio/episodes/{episode_id}/cancel:
    post:
      summary: Cancel a generating episode and reset error chunks
      description: Cancel a generating episode and reset error chunks to pending.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/episodes/{episode_id}/retry-errors:
    post:
      summary: Retry all error chunks for an episode.
      description: Retry all error chunks for an episode.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/episodes/{episode_id}/audio/{chunk_index}:
    get:
      summary: Serve a chunk's audio file.
      description: Serve a chunk's audio file.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
      - name: chunk_index
        in: path
        required: true
        schema:
          type: integer
  /api/studio/episodes/{episode_id}/audio/full:
    get:
      summary: Serve merged full episode audio.
      description: Serve merged full episode audio.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
  /api/studio/episodes/{episode_id}/move:
    put:
      summary: Move an episode to a folder.
      description: Move an episode to a folder.
      responses:
        '200':
          description: Success
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveToFolderBody'
...
info:
  title: OpenVox API
  version: 1.0.0
openapi: 3.0.3
components:
  schemas:
...
    MoveToFolderBody:
      type: object
      properties:
        folder_id:
          type: string
          description: Target folder ID, or null for root
          nullable: true
      additionalProperties: false
    CreateEpisodeBody:
      type: object
      properties:
        source_id:
          type: string
        voice_id:
          type: string
        output_format:
          type: string
        chunk_strategy:
          type: string
          description: paragraph, sentence, heading, or fixed
        chunk_max_length:
          type: integer
        code_block_rule:
          type: string
        breathing_intensity:
          type: string
          description: none, light, normal, heavy
        title:
          type: string
      required:
      - source_id
      additionalProperties: false
    UpdateEpisodeBody:
      type: object
      properties:
        title:
          type: string
      additionalProperties: false
    RegenerateWithSettingsBody:
      type: object
      properties:
        voice_id:
          type: string
        output_format:
          type: string
        chunk_strategy:
          type: string
        chunk_max_length:
          type: integer
        code_block_rule:
          type: string
        breathing_intensity:
          type: string
      additionalProperties: false
    BulkMoveEpisodesBody:
      type: object
      properties:
        episode_ids:
          type: array
          items:
            type: string
        folder_id:
          type: string
          nullable: true
      required:
      - episode_ids
      additionalProperties: false
    BulkDeleteEpisodesBody:
      type: object
      properties:
        episode_ids:
          type: array
          items:
            type: string
      required:
      - episode_ids
      additionalProperties: false
...